backend/requirements.txt
Flask
Flask-Cors
PyYAML
pandas
scikit-learn
statsmodels
matplotlib
seaborn
geopandas
rasterio
folium
rtree